+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@ Release Notes
 =============
 ## vNext
 * Az: Update `ad` commands to work with latest (breaking) structure.
+* PostgreSQL: Fix a number of issues around the introduction of Flexible Servers.
 
 ## 1.9.2
 * Container Apps: Fix to container registry credential to not emit a secret for a managed identity.