+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 Release Notes
 =============
 ## vNext
+* Az: Update `ad` commands to work with latest (breaking) structure.
 * PostgreSQL: Fix a number of issues around the introduction of Flexible Servers.
 
 ## 1.9.2